Se agrega libreria de lombok y correcciones en entidades

+++ src/main/java/py/com/fuentepy/appfinanzasBackend/repository/MonedaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import py.com.fuentepy.appfinanzasBackend.entity.Moneda;
 
-public interface MonedaRepository extends JpaRepository<Moneda, Long> {
+public interface MonedaRepository extends JpaRepository<Moneda, Integer> {
 
     Page<Moneda> findAllByOrderByIdAsc(Pageable pageable);
