+++ src/main/java/py/com/fuentepy/appfinanzasBackend/repository/TipoAhorroRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import py.com.fuentepy.appfinanzasBackend.entity.TipoAhorro;
 
-public interface TipoAhorroRepository extends JpaRepository<TipoAhorro, Long> {
+public interface TipoAhorroRepository extends JpaRepository<TipoAhorro, Integer> {
 
     Page<TipoAhorro> findAllByOrderByIdAsc(Pageable pageable);
